src/pyproject.toml
[tool.poetry]
name = "short-url"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.9"
pydantic = "^1.8.1"
click = "~7.1.2"

fastapi = "^0.63.0"
dependency-injector = "~4.36.0"
uvicorn = "^0.13.4"
psycopg2-binary = "~2.9.0"

# Formatters
openpyxl = "3.0.9"
SQLAlchemy = "1.3.12"
shortuuid = "^1.0.9"
requests = "^2.28.1"
aiofiles = "^0.8.0"
Jinja2 = "3.0.3"

[tool.poetry.dev-dependencies]
mypy = "~0.900"

# region `flake8 related`
flake8 = "~3.8.4"
bandit = "1.7.2"
wemake-python-styleguide = "~0.15.3"

isort = "~5.6.4"
black = "~20.8b1"
pytest = "~6.1.2"
coverage = "^5.5"
factory-boy = "^3.2.0"
pytest-factoryboy = "2.1.0"
pytest-mock = "^3.6.0"
pytest-unordered = "^0.4.1"
ipdb = "^0.13.9"
see = "^1.4.1"
